backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
firebase-admin==6.6.0
google-cloud-firestore>=2.19.0
google-cloud-storage==2.18.2
python-multipart==0.0.18
python-dotenv==1.0.1
google-genai>=1.0.0
google-cloud-scheduler>=2.14.0
google-cloud-run>=0.15.0
